Create initial homepage and components (mobile)
