+++ src/common.css
@@ -1,0 +1,6 @@
+@import url('https://fonts.googleapis.com/css2?family=Neucha&display=swap');
+
+.page {
+  background: url(./assets/crumpled-paper-bg.svg) no-repeat;
+  background-size: cover;
+}
