+++ src/components/Card.css
@@ -1,0 +1,4 @@
+.videoThumbnail {
+  width: 290px;
+  height: auto;
+}
