+++ tailwind.config.js
@@ -2,10 +2,10 @@ module.exports = {
   purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   darkMode: false, // or 'media' or 'class'
   theme: {
-    extend: {},
+  extend: {},
   },
   variants: {
-    extend: {},
+  extend: {},
   },
   plugins: [],
 }
